Renewed all Icons (now with authors known and English file names)
..and translated EIcons. Now the author of each icon is known, and EIcon
got a method getAuthorHtmlLink() that returns a Flaticon link. As things
are now all icons in this program are from flaticon.com and can
(legally) be used if there is a link inside the program (in the "about"
section, one for each author) and in the description of the program,
when it will be published.